Create ViewModel in onCreate instead of onAttach
This reverts a previous change that moved initialization to onAttach
from onCreate, just to have it available in a broader lifecycle.
It's expected that creating the ViewModel only in onCreate, when
saved instance state is available will be beneficial long term,
for possible integrations with saved state or Dagger Hilt.

Signed-off-by: Márton Braun <braunmarci@gmail.com>